golf/requirements.txt
flask==2.0.1
werkzeug==2.0.1
pytesseract==0.3.8
Pillow==9.5.0
opencv-python==4.7.0.72
numpy==1.24.2
pandas==1.5.3
matplotlib==3.7.1
